c3b0a8e331c1901773015038cd742701549d5d97 chore: update deps and bump MSVR to 1.88 (Cameron Garnham)

Pull request description:

  ``` bash
  bash-5.3$ cargo update --verbose
      Updating crates.io index
       Locking 83 packages to latest Rust 1.85 compatible versions
      Updating arc-swap v1.9.0 -> v1.9.1
      Updating aws-lc-rs v1.16.2 -> v1.16.3
      Updating aws-lc-sys v0.39.0 -> v0.40.0
      Updating axum v0.8.8 -> v0.8.9
      Updating bitflags v2.11.0 -> v2.11.1
        Adding block-buffer v0.12.0
     Unchanged camino v1.1.12 (available: v1.2.2)
      Updating cc v1.2.57 -> v1.2.60
      Removing chumsky v0.9.3
      Updating clap v4.6.0 -> v4.6.1
      Updating clap_derive v4.6.0 -> v4.6.1
      Updating cmake v0.1.57 -> v0.1.58
        Adding const-oid v0.10.2
      Removing crypto-common v0.1.7
        Adding crypto-common v0.1.6 (available: v0.1.7)
        Adding crypto-common v0.2.1
        Adding digest v0.11.2
      Updating env_filter v1.0.0 -> v1.0.1
      Updating env_logger v0.11.9 -> v0.11.10
      Updating fastrand v2.3.0 -> v2.4.1
      Updating fragile v2.0.1 -> v2.1.0
      Updating generic-array v0.14.7 -> v0.14.9
      Updating hashbrown v0.16.1 -> v0.17.0
     Unchanged home v0.5.11 (available: v0.5.12, requires Rust 1.88)
        Adding hybrid-array v0.4.10
      Updating hyper v1.8.1 -> v1.9.0
      Updating hyper-rustls v0.27.7 -> v0.27.9
     Unchanged icu_collections v2.1.1 (available: v2.2.0, requires Rust 1.86)
     Unchanged icu_locale_core v2.1.1 (available: v2.2.0, requires Rust 1.86)
     Unchanged icu_normalizer v2.1.1 (available: v2.2.0, requires Rust 1.86)
     Unchanged icu_normalizer_data v2.1.1 (available: v2.2.0, requires Rust 1.86)
     Unchanged icu_properties v2.1.2 (available: v2.2.0, requires Rust 1.86)
     Unchanged icu_properties_data v2.1.2 (available: v2.2.0, requires Rust 1.86)
     Unchanged icu_provider v2.1.1 (available: v2.2.0, requires Rust 1.86)
     Unchanged image v0.25.9 (available: v0.25.10, requires Rust 1.88.0)
      Updating indexmap v2.13.0 -> v2.14.0
      Updating iri-string v0.7.11 -> v0.7.12
      Updating js-sys v0.3.91 -> v0.3.95
      Updating lettre v0.11.19 -> v0.11.21
      Updating libc v0.2.183 -> v0.2.185
      Updating libredox v0.1.14 -> v0.1.16
      Updating litemap v0.8.1 -> v0.8.2
     Unchanged matchit v0.8.4 (available: v0.8.6)
      Updating mio v1.1.1 -> v1.2.0
      Updating num-conv v0.1.0 -> v0.2.1
      Updating openssl v0.10.76 -> v0.10.77
      Updating openssl-sys v0.9.112 -> v0.9.113
      Removing pin-utils v0.1.0
      Updating pkg-config v0.3.32 -> v0.3.33
      Updating potential_utf v0.1.4 -> v0.1.5
      Removing psm v0.1.24
      Removing rand v0.9.2
      Removing rand v0.10.0
        Adding rand v0.9.4
        Adding rand v0.10.1
      Updating rand_core v0.10.0 -> v0.10.1
      Updating redox_syscall v0.7.3 -> v0.7.4
      Updating rustc-hash v2.1.1 -> v2.1.2
      Updating rustls v0.23.37 -> v0.23.38
      Updating rustls-webpki v0.103.10 -> v0.103.12
      Updating semver v1.0.27 -> v1.0.28
      Updating serde_spanned v1.0.4 -> v1.1.1
     Unchanged serde_with v3.17.0 (available: v3.18.0, requires Rust 1.88)
     Unchanged serde_with_macros v3.17.0 (available: v3.18.0, requires Rust 1.88)
        Adding sha2 v0.11.0
      Updating simd-adler32 v0.3.8 -> v0.3.9
      Updating simple_asn1 v0.6.3 -> v0.6.4
      Removing stacker v0.1.23
      Updating thin-vec v0.2.14 -> v0.2.16
      Updating time v0.3.45 -> v0.3.47 (requires Rust 1.88.0)
      Updating time-core v0.1.7 -> v0.1.8 (requires Rust 1.88.0)
      Updating time-macros v0.2.25 -> v0.2.27 (requires Rust 1.88.0)
      Updating tinystr v0.8.2 -> v0.8.3
      Updating tokio v1.50.0 -> v1.52.0
      Updating tokio-macros v2.6.1 -> v2.7.0
      Updating toml v1.0.7+spec-1.1.0 -> v1.1.2+spec-1.1.0
      Updating toml_datetime v1.0.1+spec-1.1.0 -> v1.1.1+spec-1.1.0
      Updating toml_parser v1.0.10+spec-1.1.0 -> v1.1.2+spec-1.1.0
      Updating toml_writer v1.0.7+spec-1.1.0 -> v1.1.1+spec-1.1.0
      Updating unicode-segmentation v1.12.0 -> v1.13.2
      Updating uuid v1.22.0 -> v1.23.1
     Unchanged wasip2 v1.0.1+wasi-0.2.4 (available: v1.0.2+wasi-0.2.9, requires Rust 1.87.0)
     Unchanged wasip3 v0.4.0+wasi-0.3.0-rc-2026-01-06 (requires Rust 1.87.0)
      Updating wasm-bindgen v0.2.114 -> v0.2.118
      Updating wasm-bindgen-futures v0.4.64 -> v0.4.68
      Updating wasm-bindgen-macro v0.2.114 -> v0.2.118
      Updating wasm-bindgen-macro-support v0.2.114 -> v0.2.118
      Updating wasm-bindgen-shared v0.2.114 -> v0.2.118
      Updating wasm-bindgen-test v0.3.64 -> v0.3.68
      Updating wasm-bindgen-test-macro v0.3.64 -> v0.3.68
      Updating wasm-bindgen-test-shared v0.2.114 -> v0.2.118
      Updating web-sys v0.3.91 -> v0.3.95
      Updating webpki-root-certs v1.0.6 -> v1.0.7
      Updating webpki-roots v1.0.6 -> v1.0.7
      Updating winnow v1.0.0 -> v1.0.1
     Unchanged wit-bindgen v0.51.0 (requires Rust 1.87.0)
     Unchanged wit-bindgen-core v0.51.0 (requires Rust 1.87.0)
     Unchanged wit-bindgen-rust v0.51.0 (requires Rust 1.87.0)
      Updating writeable v0.6.2 -> v0.6.3
      Updating yoke v0.8.1 -> v0.8.2
      Updating yoke-derive v0.8.1 -> v0.8.2
      Updating zerocopy v0.8.47 -> v0.8.48
      Updating zerocopy-derive v0.8.47 -> v0.8.48
      Updating zerofrom v0.1.6 -> v0.1.7
      Updating zerofrom-derive v0.1.6 -> v0.1.7
      Updating zerotrie v0.2.3 -> v0.2.4
      Updating zerovec v0.11.5 -> v0.11.6
      Updating zerovec-derive v0.11.2 -> v0.11.3
  ```

  ``` bash
  bash-5.3$ cargo update --verbose
      Updating crates.io index
       Locking 16 packages to latest Rust 1.88 compatible versions
     Unchanged camino v1.1.12 (available: v1.2.2)
     Unchanged crypto-common v0.1.6 (available: v0.1.7)
      Updating darling v0.21.3 -> v0.23.0
      Updating darling_core v0.21.3 -> v0.23.0
      Updating darling_macro v0.21.3 -> v0.23.0
      Updating home v0.5.11 -> v0.5.12
      Updating icu_collections v2.1.1 -> v2.2.0
      Updating icu_locale_core v2.1.1 -> v2.2.0
      Updating icu_normalizer v2.1.1 -> v2.2.0
      Updating icu_normalizer_data v2.1.1 -> v2.2.0
      Updating icu_properties v2.1.2 -> v2.2.0
      Updating icu_properties_data v2.1.2 -> v2.2.0
      Updating icu_provider v2.1.1 -> v2.2.0
      Updating image v0.25.9 -> v0.25.10
     Unchanged matchit v0.8.4 (available: v0.8.6)
      Updating moxcms v0.7.11 -> v0.8.1
      Updating serde_with v3.17.0 -> v3.18.0
      Updating serde_with_macros v3.17.0 -> v3.18.0
      Updating wasip2 v1.0.1+wasi-0.2.4 -> v1.0.2+wasi-0.2.9
      Removing wit-bindgen v0.46.0
  ```

ACKs for top commit:
  da2ce7:
    ACK c3b0a8e331c1901773015038cd742701549d5d97

Tree-SHA512: a4c840494662d304ea8cbfc4a67bbf23203c4972fceb3116f924289e40179e934f497aaa25662cb6d8bc41c2e9a9b99af1ff518fa9155ec2c30dd48e3dd3d241

+++ src/services/user.rs
@@ -139,10 +139,10 @@ impl RegistrationService {
             None => None,
         };
 
-        if let Some(email) = &opt_email {
-            if !validate_email_address(email) {
-                return Err(UserError::EmailInvalid);
-            }
+        if let Some(email) = &opt_email
+            && !validate_email_address(email)
+        {
+            return Err(UserError::EmailInvalid);
         }
 
         validate_password_constraints(
@@ -167,19 +167,19 @@ impl RegistrationService {
             drop(self.user_repository.grant_admin_role(&user_id).await);
         }
 
-        if let Some(email) = &registration.email {
-            if email.verification_required {
-                // Email verification is enabled
-                if let Some(email) = opt_email {
-                    let mail_res = self
-                        .mailer
-                        .send_verification_mail(&email, &registration_form.username, user_id, api_base_url)
-                        .await;
+        if let Some(email) = &registration.email
+            && email.verification_required
+        {
+            // Email verification is enabled
+            if let Some(email) = opt_email {
+                let mail_res = self
+                    .mailer
+                    .send_verification_mail(&email, &registration_form.username, user_id, api_base_url)
+                    .await;
 
-                    if mail_res.is_err() {
-                        drop(self.user_repository.delete(&user_id).await);
-                        return Err(UserError::FailedToSendVerificationEmail);
-                    }
+                if mail_res.is_err() {
+                    drop(self.user_repository.delete(&user_id).await);
+                    return Err(UserError::FailedToSendVerificationEmail);
                 }
             }
         }
